Rename type_traits.hpp to TypeTraits.hpp
